server actions consolidated under a const in sdk-server
