+++ examples/react-components/src/app/dashboard/page.tsx
@@ -4,12 +4,10 @@ import {
   Export,
   Import,
   useTurnkey,
-  getSuborgs,
-  getVerifiedSuborgs,
   OtpVerification,
   OtpType,
-  initOtpAuth,
 } from "@turnkey/sdk-react";
+import { server } from "@turnkey/sdk-server";
 import { useEffect, useState } from "react";
 import "./dashboard.css";
 import {
@@ -76,6 +74,7 @@ export default function Dashboard() {
   const [emailInput, setEmailInput] = useState("");
   const [phoneInput, setPhoneInput] = useState("");
 
+
   const handleExportSuccess = async () => {
     toast.success("Wallet successfully exported");
   };
@@ -84,7 +83,7 @@ export default function Dashboard() {
     toast.success("Wallet successfully imported");
   };
   const handleResendEmail = async () => {
-    const initAuthResponse = await initOtpAuth({
+    const initAuthResponse = await server.initOtpAuth({
       suborgID: suborgId,
       otpType: OtpType.Email,
       contact: emailInput,
@@ -97,7 +96,7 @@ export default function Dashboard() {
     setOtpId(initAuthResponse?.otpId!);
   };
   const handleResendSms = async () => {
-    const initAuthResponse = await initOtpAuth({
+    const initAuthResponse = await server.initOtpAuth({
       suborgID: suborgId,
       otpType: OtpType.Sms,
       contact: phoneInput,
@@ -125,7 +124,7 @@ export default function Dashboard() {
       toast.error("Please enter a valid email address");
       return;
     }
-    const suborgs = await getVerifiedSuborgs({
+    const suborgs = await server.getVerifiedSuborgs({
       filterType: "EMAIL",
       filterValue: emailInput,
     });
@@ -139,7 +138,7 @@ export default function Dashboard() {
       userEmail: emailInput,
       userTagIds: [],
     });
-    const initAuthResponse = await initOtpAuth({
+    const initAuthResponse = await server.initOtpAuth({
       suborgID: suborgId,
       otpType: OtpType.Email,
       contact: emailInput,
@@ -159,7 +158,7 @@ export default function Dashboard() {
       toast.error("Please enter a valid phone number.");
       return;
     }
-    const suborgs = await getVerifiedSuborgs({
+    const suborgs = await server.getVerifiedSuborgs({
       filterType: "PHONE_NUMBER",
       filterValue: phoneInput,
     });
@@ -173,7 +172,7 @@ export default function Dashboard() {
       userPhoneNumber: phoneInput,
       userTagIds: [],
     });
-    const initAuthResponse = await initOtpAuth({
+    const initAuthResponse = await server.initOtpAuth({
       suborgID: suborgId,
       otpType: OtpType.Sms,
       contact: phoneInput,
@@ -223,7 +222,7 @@ export default function Dashboard() {
         console.error(`Unknown OAuth type: ${oauthType}`);
     }
     if (oidcToken) {
-      const suborgs = await getSuborgs({
+      const suborgs = await server.getSuborgs({
         filterType: "OIDC_TOKEN",
         filterValue: oidcToken.idToken,
       });
@@ -323,7 +322,7 @@ export default function Dashboard() {
           });
           setWallets(walletsResponse.wallets);
           if (userResponse.user.userEmail) {
-            const suborgs = await getVerifiedSuborgs({
+            const suborgs = await server.getVerifiedSuborgs({
               filterType: "EMAIL",
               filterValue: userResponse.user.userEmail,
             });
@@ -337,7 +336,7 @@ export default function Dashboard() {
             }
           }
           if (userResponse.user.userPhoneNumber) {
-            const suborgs = await getVerifiedSuborgs({
+            const suborgs = await server.getVerifiedSuborgs({
               filterType: "PHONE_NUMBER",
               filterValue: userResponse.user.userPhoneNumber,
             });
