--- packages/sdk-react/src/actions/index.ts
@@ -1,8 +1,0 @@
-// actions that need to be run on the server side (i.e Activities signed by the Parent Organization)
-
-export { createSuborg } from "./createSuborg";
-export { getSuborgs } from "./getSuborgs";
-export { getVerifiedSuborgs } from "./getVerifiedSuborgs";
-export { initOtpAuth } from "./initOtpAuth";
-export { oauth } from "./oauth";
-export { otpAuth } from "./otpAuth";
