+++ packages/sdk-server/package.json
@@ -48,6 +48,7 @@
     "@turnkey/api-key-stamper": "workspace:*",
     "@turnkey/http": "workspace:*",
     "buffer": "^6.0.3",
+    "next": "^15.0.2",
     "elliptic": "^6.5.5"
   },
   "devDependencies": {
