+++ packages/sdk-server/src/index.ts
@@ -31,6 +31,7 @@ import type {
 import type * as TurnkeySDKApiTypes from "./__generated__/sdk_api_types";
 
 import { fetch } from "./universal";
+import { createSuborg, getSuborgs, getVerifiedSuborgs, initOtpAuth, oauth, otpAuth } from "./actions";
 
 // Classes
 export {
@@ -40,6 +41,16 @@ export {
   TurnkeyServerSDK as Turnkey,
   TurnkeyServerClient,
   TurnkeyRequestError,
+};
+
+// Server Actions
+export const server = {
+  initOtpAuth,
+  otpAuth,
+  oauth,
+  getSuborgs,
+  getVerifiedSuborgs,
+  createSuborg,
 };
 
 // Types
